Rearranges test stuff.
