+++ lib/parser/test-utils.ts
@@ -1,3 +1,4 @@
+import { getEntityManager } from 'lib/db/test-utils';
 import type { DeepPartial, EntityManager } from 'typeorm';
 import type { Claim, Import } from 'lib/db/entities';
 
@@ -44,6 +45,25 @@ export function parseTestClaim( rawClaim: RawTestClaim ): DeepPartial< Claim > {
 		cost: Number.parseFloat( rawClaim.cost ),
 		provider: undefined,
 	};
+}
+
+export const baseClaim = {
+	number: '1234',
+	status: 'pending',
+	type: 'test',
+	slug: 'test',
+	serviceDate: new Date( 2021, 0, 1, 12 ),
+	billed: 1.23,
+	cost: 1.23,
+} as const;
+
+export async function getEntityManagerWithClaim(): Promise< EntityManager > {
+	const em = getEntityManager();
+	await em.insert( 'Claim', {
+		...baseClaim,
+		slug: 'test1',
+	} );
+	return em;
 }
 
 export async function getImportEntity( em: EntityManager ): Promise< Import > {
